test(conversations): consolidate bbolt conversation store integration tests
